+++ src/mpol/__init__.py
@@ -1,2 +1,1 @@
-from mpol.mpol_version import __version__
 zenodo_record = 10064221