+++ test/fourier_test.py
@@ -1,9 +1,8 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import torch
+from mpol import fourier, utils
 from pytest import approx
-
-from mpol import fourier, images, utils
 
 
 def test_fourier_cube(coords, tmp_path):
@@ -142,17 +141,14 @@ def test_predict_vis_nufft(coords, baselines_1D):
 
     nchan = 10
 
-    # instantiate an BaseCube layer filled with zeros
-    basecube = images.BaseCube(coords=coords, nchan=nchan, pixel_mapping=lambda x: x)
-    imagecube = images.ImageCube(coords=coords, nchan=nchan)
-
     # we have a multi-channel cube, but only sent single-channel uu and vv
     # coordinates. The expectation is that TorchKbNufft will parallelize these
 
     layer = fourier.NuFFT(coords=coords, nchan=nchan)
 
     # predict the values of the cube at the u,v locations
-    output = layer(imagecube(basecube()), uu, vv)
+    blank_packed_img = torch.zeros((nchan, coords.npix, coords.npix))
+    output = layer(blank_packed_img, uu, vv)
 
     # make sure we got back the number of visibilities we expected
     assert output.shape == (nchan, len(uu))
@@ -172,62 +168,55 @@ def test_predict_vis_nufft_cached(coords, baselines_1D):
 
     nchan = 10
 
-    # instantiate an ImageCube layer filled with zeros
-    # instantiate an BaseCube layer filled with zeros
-    basecube = images.BaseCube(coords=coords, nchan=nchan, pixel_mapping=lambda x: x)
-    imagecube = images.ImageCube(coords=coords, nchan=nchan)
-
     # we have a multi-channel cube, but sent only single-channel uu and vv
     # coordinates. The expectation is that TorchKbNufft will parallelize these
-
     layer = fourier.NuFFTCached(coords=coords, nchan=nchan, uu=uu, vv=vv)
 
     # predict the values of the cube at the u,v locations
-    output = layer(imagecube(basecube()))
+    blank_packed_img = torch.zeros((nchan, coords.npix, coords.npix))
+    output = layer(blank_packed_img)
 
     # make sure we got back the number of visibilities we expected
     assert output.shape == (nchan, len(uu))
 
     # if the image cube was filled with zeros, then we should make sure this is true
     assert output.detach().numpy() == approx(
-        np.zeros((nchan, len(uu)), dtype=np.complex128)
+        np.zeros((nchan, len(uu)))
     )
 
 
 def test_nufft_cached_predict_GPU(coords, baselines_1D):
-    if not torch.cuda.is_available():
-        pass
+    if torch.cuda.is_available():
+        device = torch.device("cuda")
     else:
-        device = torch.device("cuda:0")
+        return
 
-        # just see that we can load the layer and get something through without error
-        # for a very simple blank function
+    # just see that we can load the layer and get something through without error
+    # for a very simple blank function
 
-        # load some data
-        uu, vv = baselines_1D
+    # load some data
+    uu, vv = baselines_1D
 
-        nchan = 10
+    nchan = 10
 
-        # instantiate an ImageCube layer filled with zeros and send to GPU
-        imagecube = images.ImageCube(coords=coords, nchan=nchan).to(device=device)
+    # we have a multi-channel cube, but only sent single-channel uu and vv
+    # coordinates. The expectation is that TorchKbNufft will parallelize these
 
-        # we have a multi-channel cube, but only sent single-channel uu and vv
-        # coordinates. The expectation is that TorchKbNufft will parallelize these
+    layer = fourier.NuFFTCached(coords=coords, nchan=nchan, uu=uu, vv=vv).to(
+        device=device
+    )
 
-        layer = fourier.NuFFTCached(coords=coords, nchan=nchan, uu=uu, vv=vv).to(
-            device=device
-        )
+    # predict the values of the cube at the u,v locations
+    blank_packed_img = torch.zeros((nchan, coords.npix, coords.npix)).to(device=device)
+    output = layer(blank_packed_img)
 
-        # predict the values of the cube at the u,v locations
-        output = layer(imagecube())
+    # make sure we got back the number of visibilities we expected
+    assert output.shape == (nchan, len(uu))
 
-        # make sure we got back the number of visibilities we expected
-        assert output.shape == (nchan, len(uu))
-
-        # if the image cube was filled with zeros, then we should make sure this is true
-        assert output.cpu().detach().numpy() == approx(
-            np.zeros((nchan, len(uu)), dtype=np.complex128)
-        )
+    # if the image cube was filled with zeros, then we should make sure this is true
+    assert output.cpu().detach().numpy() == approx(
+        np.zeros((nchan, len(uu)), dtype=np.complex128)
+    )
 
 
 def test_nufft_accuracy_single_chan(coords, baselines_1D, tmp_path):
@@ -325,12 +314,10 @@ def test_nufft_cached_accuracy_single_chan(coords, baselines_1D, tmp_path):
     img_packed = utils.sky_gaussian_arcsec(
         coords.packed_x_centers_2D, coords.packed_y_centers_2D, **kw
     )
-    img_packed_tensor = torch.tensor(img_packed[np.newaxis, :, :], requires_grad=True)
+    img_packed_tensor = torch.tensor(img_packed[np.newaxis, :, :], requires_grad=True, dtype=torch.float32)
 
     # use the NuFFT to predict the values of the cube at the u,v locations
-    num_output = (
-        layer(img_packed_tensor)[0]
-    )  # take the channel dim out
+    num_output = layer(img_packed_tensor)[0]  # take the channel dim out
 
     # calculate the values analytically
     an_output = utils.fourier_gaussian_lambda_arcsec(uu, vv, **kw)
@@ -340,7 +327,6 @@ def test_nufft_cached_accuracy_single_chan(coords, baselines_1D, tmp_path):
     max_diff = torch.max(torch.abs(diff))
     max = torch.max(torch.abs(num_output))
     print(max_diff, max)
-
 
     # collapse the function into 1D by doing q
     qq = utils.torch2npy(torch.hypot(uu, vv))
@@ -404,7 +390,7 @@ def test_nufft_cached_accuracy_coil_broadcast(coords, baselines_1D):
     # broadcast to 5 channels -- the image will be the same for each
     img_packed_tensor = torch.tensor(
         img_packed[np.newaxis, :, :] * np.ones((nchan, coords.npix, coords.npix)),
-        requires_grad=True,
+        requires_grad=True, dtype=torch.float32
     )
 
     # use the NuFFT to predict the values of the cube at the u,v locations