+++ test/geometry_test.py
@@ -1,13 +1,13 @@
+import numpy as np
 import torch
-from pytest import approx
-import numpy as np 
-
 from mpol import geometry
+from pytest import approx
+
 
 def test_rotate_points():
-    '''
+    """
     Test rotation from flat 2D frame to observer frame and back
-    '''
+    """
     xs = torch.tensor([0.0, 1.0, 2.0])
     ys = torch.tensor([1.0, -1.0, 2.0])
 
@@ -29,14 +29,13 @@ def test_rotate_points():
 
 
 def test_rotate_coords(coords):
-    
+
     omega = 35. * np.pi/180
     incl = 30. * np.pi/180
     Omega = 210. * np.pi/180
 
     x, y = geometry.observer_to_flat(coords.sky_x_centers_2D, coords.sky_y_centers_2D, omega=omega, incl=incl, Omega=Omega)
-    
+
     print(x, y)
 
-    
-    
+