+++ test/utils_test.py
@@ -1,9 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
-
 from mpol import coordinates, utils
-from mpol.constants import *
 
 
 @pytest.fixture